Update and rename test_wu2004.py to test_util.py
Removed lines 6-10, 22-26, 43-46
Renamed the file test_wu2004.py and updated to test_util.py
Inside test_util.py add from emll.util import series_to_hdi
In the beginning of wu2004.ipynb add from emll.util import series_to_hdi
Removed series_to_hdi function (lines 13–20).
